--- src/components/sidenav/styles/sidebar-option.scss
@@ -1,134 +1,0 @@
-@import 'src/css/quasar.variables.scss';
-@import 'src/css/utils.scss';
-
-.wrapper {
-  background: $gray-1;
-  border-bottom: 1px solid $object-light;
-  box-shadow: 0px 0px 24px 4px rgba(0, 0, 0, 0.08);
-  padding-top: 16px;
-  overflow-x: hidden;
-}
-
-.q-expansion-item {
-  padding: 8px 16px;
-}
-
-.q-item__section {
-  min-width: 0px !important;
-}
-
-.exansion-title {
-  margin-left: 8px;
-  font-weight: 400;
-  font-size: 16px;
-  line-height: 19px;
-}
-
-.active-item {
-  border: 1px solid $astar-blue;
-}
-
-.expansion-list {
-  width: 100%;
-  padding-left: 24px;
-  padding-right: 18px;
-}
-
-.ic-item {
-  width: 100%;
-  background: #fff;
-  margin: 16px;
-  min-height: 40px;
-  border-radius: 6px;
-  &:focus {
-    border: 1px solid $astar-blue;
-  }
-
-  .icon {
-    margin-right: 8px;
-  }
-}
-
-.header-icon {
-  width: 20px;
-  color: $navy-1;
-}
-
-.icon {
-  width: 20px;
-  color: $astar-blue;
-}
-
-.item-name {
-  font-weight: 500;
-  font-size: 14px;
-  line-height: 17px;
-  color: $navy-1;
-}
-
-.wrapper--option {
-  .q-expansion-item {
-    padding: 0 !important;
-    margin-left: 16px;
-    margin-top: 16px;
-    margin-bottom: 16px;
-
-    &:focus {
-      border: 1px solid $astar-blue;
-    }
-  }
-
-  .wrapper--sub-item {
-    width: 100%;
-    padding-left: 16px;
-  }
-
-  .ic-item-lang-active {
-    border: 1px solid $astar-blue;
-  }
-
-  .item--sub {
-    margin: 0px;
-    margin-top: 16px;
-  }
-}
-
-.button--close {
-  font-weight: 500;
-  font-size: 14px;
-  line-height: 17px;
-  display: flex;
-  margin: 16px;
-  padding-bottom: 40px;
-  justify-content: center;
-  color: $navy-1;
-  cursor: pointer;
-}
-
-.body--dark {
-  .wrapper {
-    background: $body-bg-dark;
-    box-shadow: 0px 2px 2px rgba(0, 0, 0, 0.25);
-    border-bottom: 2px solid $gray-5;
-  }
-
-  .active-item {
-    border: 1px solid $astar-blue;
-  }
-
-  .ic-item {
-    background: rgba(255, 255, 255, 0.05);
-  }
-
-  .item-name {
-    color: $gray-1;
-  }
-
-  .header-icon {
-    color: #fff;
-  }
-
-  .button--close {
-    color: $gray-1;
-  }
-}
